created an 'abstract' class Client, with the child classes ApiKeySendClient and EmailClient
