+++ sendgrid.go
@@ -50,11 +50,6 @@ func createSendGridRequest(sgOptions sendGridOptions) rest.Request {
 	return requestNew(options)
 }
 
-// NewSendClient constructs a new Twilio SendGrid client given an API key
-func NewSendClient(key string) *Client {
-	return &Client{apiKey: key}
-}
-
 // extractEndpoint extracts the endpoint from a baseURL
 func extractEndpoint(link string) (string, error) {
 	parsedURL, err := url.Parse(link)
