+++ twilio_email.go
@@ -1,10 +1,18 @@
 package sendgrid
 
 import (
+	"bytes"
+	"compress/gzip"
 	"encoding/base64"
 
 	"github.com/sendgrid/rest"
+	"github.com/sendgrid/sendgrid-go/helpers/mail"
 )
+
+type EmailClient struct {
+	Client
+	emailOptions TwilioEmailOptions
+}
 
 // TwilioEmailOptions for GetTwilioEmailRequest
 type TwilioEmailOptions struct {
@@ -15,8 +23,8 @@ type TwilioEmailOptions struct {
 }
 
 // NewTwilioEmailSendClient constructs a new Twilio Email client given a username and password
-func NewTwilioEmailSendClient(username, password string) *Client {
-	return &Client{emailOptions: TwilioEmailOptions{Username: username, Password: password}, apiKey: ""}
+func NewTwilioEmailSendClient(username, password string) *EmailClient {
+	return &EmailClient{emailOptions: TwilioEmailOptions{Username: username, Password: password}}
 }
 
 // GetTwilioEmailRequest create Request
@@ -36,4 +44,36 @@ func GetTwilioEmailRequest(twilioEmailOptions TwilioEmailOptions) rest.Request {
 	}
 
 	return requestNew(options)
+}
+
+// prepareRequest prepares the email request with the given headers
+func (cl *EmailClient) prepareRequest(email *mail.SGMailV3, headers map[string]string) (rest.Request, error) {
+	var request rest.Request
+	request = GetTwilioEmailRequest(cl.emailOptions)
+	request.Method = "POST"
+
+	for k, v := range headers {
+		request.Headers[k] = v
+	}
+
+	request.Body = mail.GetRequestBody(email)
+	// when Content-Encoding header is set to "gzip"
+	// mail body is compressed using gzip according to
+	// https://docs.sendgrid.com/api-reference/mail-send/mail-send#mail-body-compression
+	if request.Headers["Content-Encoding"] == "gzip" {
+		var gzipped bytes.Buffer
+		gz := gzip.NewWriter(&gzipped)
+		if _, err := gz.Write(request.Body); err != nil {
+			return request, err
+		}
+		if err := gz.Flush(); err != nil {
+			return request, err
+		}
+		if err := gz.Close(); err != nil {
+			return request, err
+		}
+		request.Body = gzipped.Bytes()
+	}
+
+	return request, nil
 }
