+++ twilio_email_test.go
@@ -7,15 +7,18 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/sendgrid/sendgrid-go/helpers/mail"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-// func TestNewTwilioEmailSendClient(t *testing.T) {
-// 	mailClient := NewTwilioEmailSendClient("username", "password")
-// 	assert.Equal(t, "https://email.twilio.com/v3/mail/send", mailClient.request.BaseURL)
-// 	assert.Equal(t, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=", mailClient.request.Headers["Authorization"])
-// }
+func TestNewTwilioEmailSendClient(t *testing.T) {
+	mailClient := NewTwilioEmailSendClient("username", "password")
+	req, err := mailClient.prepareRequest(&mail.SGMailV3{}, nil)
+	assert.NoError(t, err)
+	assert.Equal(t, "https://email.twilio.com/v3/mail/send", req.BaseURL)
+	assert.Equal(t, "Basic dXNlcm5hbWU6cGFzc3dvcmQ=", req.Headers["Authorization"])
+}
 
 func TestGetTwilioEmailRequest(t *testing.T) {
 	fakeServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
